fix: create wallet at login state, update appData accordingly, remove Wallet object in appData
