+++ app/_layout.tsx
@@ -21,7 +21,7 @@ import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import CustomDrawerContent from '~/components/CustomDrawerContent';
 import { AppDataProvider } from '~/components/Wrappers/AppData';
 import { ThemeWrapper } from '~/components/Wrappers/ThemeWrapper';
-import { UpdatePrivyData } from '~/services/updateAppData';
+
 const Layout = () => {
   const [fontsLoaded] = useFonts({
     DMSerifText_400Regular,
@@ -45,7 +45,6 @@ const Layout = () => {
     <PrivyProvider appId="clxd5oc5m007jrpv8y8clt6z7">
       <ThemeWrapper>
         <AppDataProvider>
-          <UpdatePrivyData />
           <GestureHandlerRootView className="flex-1">
             <Drawer
               drawerContent={CustomDrawerContent}
