+++ services/Supabase/useSupabaseUser.ts
@@ -32,7 +32,7 @@ export const useSupabaseUser = ({ address }: SupabaseUserProps) => {
         setUser(initialData);
 
         const channel = supabase
-          .channel(`Users:address:${address}`)
+          .channel(`user:${address}`)
           .on(
             'postgres_changes',
             {
