+++ types/appData.ts
@@ -1,4 +1,4 @@
-import { PrivyEmbeddedWalletProvider, EmbeddedWalletState } from '@privy-io/expo';
+import { EmbeddedWalletState } from '@privy-io/expo';
 import * as _privy_io_public_api from '@privy-io/public-api';
 
 type User = {
@@ -6,16 +6,9 @@ type User = {
   privyID: string;
 };
 
-type Wallet = {
-  account?: _privy_io_public_api.PrivyEthereumEmbeddedWalletAccount;
-  provider?: PrivyEmbeddedWalletProvider;
-  status?: string;
-};
-
 type Privy = {
   user?: _privy_io_public_api.PrivyUser;
   wallet?: EmbeddedWalletState;
-  provider?: PrivyEmbeddedWalletProvider;
 };
 
-export type { User, Wallet, Privy };
+export type { User, Privy };
